modularized logger and added sample use
